+++ test/diffcrypt/rails/encrypted_configuration_test.rb
@@ -17,10 +17,10 @@ class Diffcrypt::Rails::EncryptedConfigurationTest < Minitest::Test
   # This verifies that encrypted and unecrypted data can't be accidently the
   # same, which would create false positive tests and a major security issue
   def test_that_fixtures_are_different
-    refute_equal File.read("#{__dir__}/../../fixtures/example.yml.enc"), File.read("#{__dir__}/../../fixtures/example.yml")
+    refute_equal ::File.read("#{__dir__}/../../fixtures/example.yml.enc"), ::File.read("#{__dir__}/../../fixtures/example.yml")
   end
 
   def test_that_fixture_can_be_decrypted
-    assert_equal configuration.read, File.read("#{__dir__}/../../fixtures/example.yml")
+    assert_equal configuration.read, ::File.read("#{__dir__}/../../fixtures/example.yml")
   end
 end